+++ resources/web/flow/src/mocks/mock.ts
@@ -1,3 +1,46 @@
+const flowDefinition = `
+version: v1beta1
+type: flow
+steps:
+- type: pass    # pass step 用于传递或转换数据，也可以作为 task step 调试的占位步骤
+  name: PreProcess
+- type: parallel
+  name: ParallelProcess
+  branches:    # parallel step 并行执行 branches 里面的步骤
+    - steps:    # 第一个 branch
+        - type: pass
+          name: WorkerStep1
+    - steps:    # 第二个 branch
+        - type: pass
+          name: WorkerStep2-1
+        - type: pass
+          name: WorkerStep2-2
+  outputMappings:    # 该 mapping 产生 {"key1":"value1", "key2":"value2"} 作为 step 输出 (output)
+    - target: key1
+      source: value1
+    - target: key2
+      source: value2
+- type: wait    # wait step 等待指定的时间，duration 单位为秒
+  name: Wait10s
+  duration: 10
+- type: choice    # choice step 根据 step input 和指定条件决定下一步骤
+  name: CheckStatus
+  choices:    # 假设该 execution 的输入为 {"key":"process_failed"}, 则该 step 输入为 {"key1":"value1", "key2":"value2", "key":"process_failed"}
+    - condition: $.key1==$.key2    # key1 和 key2 值如果相等，执行 OnSuccess 步骤，结束后进入 ProcessSucceeded 步骤
+      steps:
+        - type: pass
+          name: OnSucceess
+      goto: ProcessSucceeded
+    - condition: $.key=="process_failed"    # 如果输入中的 key 字段值为字符串 "process_failed", 跳转到 ProcessFailed 步骤
+      goto: ProcessFailed
+  default:
+      goto: ProcessSucceeded
+- type: succeed    # succeed step 是一种结束步骤，产生 ExecutionSucceeded 事件
+  name: ProcessSucceeded
+- type: fail    # fail step 是一种结束步骤，产生 ExecutionFailed 事件
+  name: ProcessFailed
+`;
+
 export function getFakeResultSet(): { [s: string]: any } {
   return {
     'describeFlow': {
@@ -99,6 +142,7 @@ steps:
     },
     'describeExecution': {
       FlowName: 'flow',
+      FlowDefinition: flowDefinition,
       Input: '{"key":"value"}',
       Name: 'exec',
       Output: '{}',
@@ -118,50 +162,10 @@ steps:
       ],
     },
     'describeFlowDefinition': {
-      Definition: `
-version: v1beta1
-type: flow
-steps:
-- type: pass    # pass step 用于传递或转换数据，也可以作为 task step 调试的占位步骤
-  name: PreProcess
-- type: parallel
-  name: ParallelProcess
-  branches:    # parallel step 并行执行 branches 里面的步骤
-    - steps:    # 第一个 branch
-        - type: pass
-          name: WorkerStep1
-    - steps:    # 第二个 branch
-        - type: pass
-          name: WorkerStep2-1
-        - type: pass
-          name: WorkerStep2-2
-  outputMappings:    # 该 mapping 产生 {"key1":"value1", "key2":"value2"} 作为 step 输出 (output)
-    - target: key1
-      source: value1
-    - target: key2
-      source: value2
-- type: wait    # wait step 等待指定的时间，duration 单位为秒
-  name: Wait10s
-  duration: 10
-- type: choice    # choice step 根据 step input 和指定条件决定下一步骤
-  name: CheckStatus
-  choices:    # 假设该 execution 的输入为 {"key":"process_failed"}, 则该 step 输入为 {"key1":"value1", "key2":"value2", "key":"process_failed"}
-    - condition: $.key1==$.key2    # key1 和 key2 值如果相等，执行 OnSuccess 步骤，结束后进入 ProcessSucceeded 步骤
-      steps:
-        - type: pass
-          name: OnSucceess
-      goto: ProcessSucceeded
-    - condition: $.key=="process_failed"    # 如果输入中的 key 字段值为字符串 "process_failed", 跳转到 ProcessFailed 步骤
-      goto: ProcessFailed
-  default:
-      goto: ProcessSucceeded
-- type: succeed    # succeed step 是一种结束步骤，产生 ExecutionSucceeded 事件
-  name: ProcessSucceeded
-- type: fail    # fail step 是一种结束步骤，产生 ExecutionFailed 事件
-`,
+      Definition: flowDefinition,
     },
     'describeInitialEntry': {
-      entry: 'definition'
+      entry: '/'
     }
   }
 }
